Move metaboxes definition for selective execution; refactoring
Move these out of functions.php because they are only executed in post editor and plugin config.

Rename "metaboxes" to "elements" because affected UI are not all metaboxes.

Wrap in a class for more elegant function names and calls.

Add another dimension to elements arrays to organize by section (corresponding changes to UI implementation in subsequent commit).  This makes it much easier to visualize how the config checkboxes relate to editor elements.  Re-ordered several elements to take advantage of this organization.

Do not use translated captions as array keys because those keys are for storage and retrieval of restriction settings.  The restrictions could only be stored for one language and would be ineffective for other languages.  Instead, the element ID or class serves as array key.  In cases where more than one item is affected, an "elements" key is added to the array for execution time use instead of the array key.  This retains the flexibility of handling future UI elements with existing restriction settings.

Function name prefixes were inconsistent: pp_capability, pp_capabilities or pp_cabapbility.  This plugin has a lot of old code and should be refactored with a better object model, but for new code let's at least use:

* pp_capabilities_ for bare function names, filter and actions
* PP_Capabilities_ for class names
* Camel case for class methods and properties

This functionality is now available for 'post' and 'page' types by default.  Additional post types can be enabled through filter  'pp_capabilities_feature_post_types' with an additional checkbox column added to the configuration screen for each post type.  See subsequent commits for this and other related changes.