Clarify naming, implementation of function checking for Classic Editor availability
For the purpose of the plugin configuration screen, we are checking whether the Classic Editor is available (even if not enabled for the current user).

Classic Editor can be made available by a number of triggers. The function has been modified to consider several of the most common (not just checking for activation of the Classic Editor plugin with standard folder name).