Move restriction application code out of main functions.php
This code is only executed for the post editor, so we don't need it to be in a module that loads on every site access.

Instead of an empty check and ternary operator on get_option() return value, just pass an empty array as the default value.

Use short syntax to create an array.  There is old code in this plugin, but we are standardizing on [] for new and modified code.

array_key_exists() returns true for null values, so it's usually best to use isset() or !empty()

Simplified post type checking by moving frequently used snippets into pp_capabilities_get_post_type()

Replaced "ce_" prefix with "classic_" for consistency