Unique Capablilities sidebar: caption for "Create" capability switch
Adjust this caption because when it's read without the tooltip, the current caption "Enable the Create column" implies this is only about enabling UI within the Role Capabilities editing screen.

This switch causes standard roles like Editor and Author to no longer be able to create posts.  So it's important for the Administrator to understand that the newly available Create column is a requirement that needs to be filled, not just an optional extra assignment.